+++ lib/pid/pid.h
@@ -9,13 +9,13 @@ class PID {
         Servo servoZ;
         int pinX;
         int pinZ;
-        float rotX; // deg/s
-        float rotZ; // deg/s
-        float angX; // deg
-        float angY; //radians
-        float angZ; // deg
-        float intX; // deg*s
-        float intZ; // deg*s
+        double rotX; // deg/s
+        double rotZ; // deg/s
+        double angX; // deg
+        double angY; //radians
+        double angZ; // deg
+        double intX; // deg*s
+        double intZ; // deg*s
         long tstep; // s
         int posX; //us - real X servo position
         int posZ; //us - real Z servo position
@@ -23,19 +23,19 @@ class PID {
         int tPosZ; //us - "true" Z servo position
         int range;
         double mag;
-        float kPX;
-        float kIX;
-        float kDX;
-        float kPZ;
-        float kIZ;
-        float kDZ;
+        double kPX;
+        double kIX;
+        double kDX;
+        double kPZ;
+        double kIZ;
+        double kDZ;
 
     public:
         PID(Servo sX, Servo sZ, int pX, int pZ);
-        float getTPosX();
-        float getTPosZ();
-        float getPosX();
-        float getPosZ();
-        void control(float aY, float rX, float rZ, float aX, float aZ, float iX, float iZ);
+        double getTPosX();
+        double getTPosZ();
+        double getPosX();
+        double getPosZ();
+        void control(double aY, double rX, double rZ, double aX, double aZ, double iX, double iZ);
 };
 #endif
